pyproject.toml
[project]
name = "lunii-cyoa"
version = "0.1.0"
description = "Pipeline to author CYOA stories, validate TOML, and compile to Lunii-ready packs."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "tomlkit>=0.12.0",
  "pydantic>=2.7.0,<3",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "pillow>=11.2.1,<12.0",
  "ffmpeg-python>=0.2.0",
  "pydub>=0.25.0",
  "tqdm>=4.66.0",
  "elevenlabs>=1.9.0",
  "python-dotenv>=1.0.1",
  "lunii-packs",
  "pipelex>=0.15.7",
  "google-genai>=1.52.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "black>=23.0.0",
  "ruff>=0.6.0",
  "mypy>=1.0.0",
  "types-requests",
  "types-toml",
]

[tool.black]
line-length = 200
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
exclude = [
  ".git",
  ".ruff_cache",
  ".mypy_cache",
  ".venv",
  "build",
  "dist",
  "__pycache__",
]
line-length = 200
indent-width = 4
target-version = "py313"

[tool.uv.sources]
lunii-packs = { path = "third_party/lunii_packs-0.0.1-py3-none-any.whl" }

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
